+++ _modules/events.js
@@ -1,0 +1,3 @@
+import Events from 'events';
+let events = new Events.EventEmitter();
+export default events;